+++ src/lib/components/index.ts
@@ -19,6 +19,7 @@ export { default as CodeBlock } from './CodeBlock/CodeBlock.svelte';
 
 export { default as ColorPicker } from './ColorPicker/ColorPicker.svelte';
 
+export { default as DarkProvider } from './Dark/DarkProvider.svelte';
 export { default as DarkToggle } from './Dark/DarkToggle.svelte';
 
 export { default as Dropdown } from './Dropdown/Dropdown.svelte';